Refactored enum into seperated header
