+++ src/utility/enums.hpp
@@ -1,0 +1,12 @@
+#ifndef SRC_UTILITY_ENUMS_H
+#define SRC_UTILITY_ENUMS_H
+
+enum class BloodGroup{
+  APositive, ANegative,
+  BPositive, BNegative,
+  OPositive, ONegative,
+  ABPositive, ABNegative
+};
+
+
+#endif
